Move retry logic from adapter to call sites in apply.go and bundle_plan.go
Adapter is a type-adaptation layer; retries are an operational concern.
- New bundle/direct/retry.go: isTransient, retryWith/retryOnTransient/retryErr
- dresources/retry.go: only retrySafe/IsRetrySafe/UnwrapRetrySafe (opt-in signal)
- adapter.go: stripped to pure type adaptation
- apply.go, bundle_plan.go: retries applied at each adapter call site
