+++ bundle/direct/apply.go
@@ -8,6 +8,7 @@ import (
 	"reflect"
 
 	"github.com/databricks/cli/bundle/deployplan"
+	"github.com/databricks/cli/bundle/direct/dresources"
 	"github.com/databricks/cli/bundle/direct/dstate"
 	"github.com/databricks/cli/libs/log"
 	"github.com/databricks/databricks-sdk-go/apierr"
@@ -50,7 +51,16 @@ func (d *DeploymentUnit) Deploy(ctx context.Context, db *dstate.DeploymentState,
 }
 
 func (d *DeploymentUnit) Create(ctx context.Context, db *dstate.DeploymentState, newState any) error {
-	newID, remoteState, err := d.Adapter.DoCreate(ctx, newState)
+	var newID string
+	var remoteState any
+	_, err := retryWith(ctx, func(err error) bool {
+		return dresources.IsRetrySafe(err) && isTransient(ctx, err)
+	}, func() (struct{}, error) {
+		var e error
+		newID, remoteState, e = d.Adapter.DoCreate(ctx, newState)
+		return struct{}{}, e
+	})
+	err = dresources.UnwrapRetrySafe(err)
 	if err != nil {
 		// No need to prefix error, there is no ambiguity (only one operation - DoCreate) and no additional context (like id)
 		return err
@@ -68,7 +78,9 @@ func (d *DeploymentUnit) Create(ctx context.Context, db *dstate.DeploymentState,
 		return fmt.Errorf("saving state after creating id=%s: %w", newID, err)
 	}
 
-	waitRemoteState, err := d.Adapter.WaitAfterCreate(ctx, newID, newState)
+	waitRemoteState, err := retryOnTransient(ctx, func() (any, error) {
+		return d.Adapter.WaitAfterCreate(ctx, newID, newState)
+	})
 	if err != nil {
 		return fmt.Errorf("waiting after creating id=%s: %w", newID, err)
 	}
@@ -91,7 +103,7 @@ func (d *DeploymentUnit) Recreate(ctx context.Context, db *dstate.DeploymentStat
 	// MANAGED_BY_PARENT is still disregarded — the subsequent Create with
 	// replace_existing=true will reconfigure the parent-managed resource in
 	// place, matching the Terraform provider's recreate behaviour.
-	err = d.Adapter.DoDelete(ctx, oldID, oldState)
+	err = retryErr(ctx, func() error { return d.Adapter.DoDelete(ctx, oldID, oldState) })
 	if err != nil && !isResourceGone(err) && !isManagedByParent(err) {
 		return fmt.Errorf("deleting old id=%s: %w", oldID, err)
 	}
@@ -120,7 +132,9 @@ func (d *DeploymentUnit) Update(ctx context.Context, db *dstate.DeploymentState,
 		return fmt.Errorf("internal error: DoUpdate not implemented for resource %s", d.ResourceKey)
 	}
 
-	remoteState, err := d.Adapter.DoUpdate(ctx, id, newState, planEntry)
+	remoteState, err := retryOnTransient(ctx, func() (any, error) {
+		return d.Adapter.DoUpdate(ctx, id, newState, planEntry)
+	})
 	if err != nil {
 		return fmt.Errorf("updating id=%s: %w", id, err)
 	}
@@ -135,7 +149,9 @@ func (d *DeploymentUnit) Update(ctx context.Context, db *dstate.DeploymentState,
 		return fmt.Errorf("saving state id=%s: %w", id, err)
 	}
 
-	waitRemoteState, err := d.Adapter.WaitAfterUpdate(ctx, id, newState)
+	waitRemoteState, err := retryOnTransient(ctx, func() (any, error) {
+		return d.Adapter.WaitAfterUpdate(ctx, id, newState)
+	})
 	if err != nil {
 		return fmt.Errorf("waiting after updating id=%s: %w", id, err)
 	}
@@ -150,7 +166,13 @@ func (d *DeploymentUnit) Update(ctx context.Context, db *dstate.DeploymentState,
 }
 
 func (d *DeploymentUnit) UpdateWithID(ctx context.Context, db *dstate.DeploymentState, oldID string, newState any) error {
-	newID, remoteState, err := d.Adapter.DoUpdateWithID(ctx, oldID, newState)
+	var newID string
+	var remoteState any
+	err := retryErr(ctx, func() error {
+		var e error
+		newID, remoteState, e = d.Adapter.DoUpdateWithID(ctx, oldID, newState)
+		return e
+	})
 	if err != nil {
 		return fmt.Errorf("updating id=%s: %w", oldID, err)
 	}
@@ -171,7 +193,9 @@ func (d *DeploymentUnit) UpdateWithID(ctx context.Context, db *dstate.Deployment
 		return fmt.Errorf("saving state id=%s: %w", oldID, err)
 	}
 
-	waitRemoteState, err := d.Adapter.WaitAfterUpdate(ctx, newID, newState)
+	waitRemoteState, err := retryOnTransient(ctx, func() (any, error) {
+		return d.Adapter.WaitAfterUpdate(ctx, newID, newState)
+	})
 	if err != nil {
 		return fmt.Errorf("waiting after updating id=%s: %w", newID, err)
 	}
@@ -221,7 +245,7 @@ func (d *DeploymentUnit) Delete(ctx context.Context, db *dstate.DeploymentState,
 }
 
 func (d *DeploymentUnit) Resize(ctx context.Context, db *dstate.DeploymentState, id string, newState any) error {
-	err := d.Adapter.DoResize(ctx, id, newState)
+	err := retryErr(ctx, func() error { return d.Adapter.DoResize(ctx, id, newState) })
 	if err != nil {
 		return fmt.Errorf("resizing id=%s: %w", id, err)
 	}
@@ -265,7 +289,9 @@ func (d *DeploymentUnit) refreshRemoteState(ctx context.Context, id string) erro
 	if d.RemoteState != nil {
 		return nil
 	}
-	remoteState, err := d.Adapter.DoRead(ctx, id)
+	remoteState, err := retryOnTransient(ctx, func() (any, error) {
+		return d.Adapter.DoRead(ctx, id)
+	})
 	if err != nil {
 		return fmt.Errorf("failed to refresh remote state id=%s: %w", id, err)
 	}
